+++ src/DateTimeZone.php
@@ -3,7 +3,7 @@
  * Date time zone
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2019 Pronamic
+ * @copyright 2005-2020 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\DateTime
  * @see       https://github.com/woocommerce/woocommerce/blob/3.3.4/includes/class-wc-datetime.php
@@ -16,8 +16,9 @@ namespace Pronamic\WordPress\DateTime;
  * Date time zone
  *
  * @author  Remco Tolsma
- * @version 1.0.1
+ * @version 1.2.0
  * @since   1.0.0
+ * @psalm-immutable
  */
 class DateTimeZone extends \DateTimeZone {
 	/**
@@ -26,46 +27,45 @@ class DateTimeZone extends \DateTimeZone {
 	 * @link https://github.com/Rarst/wpdatetime/blob/0.3/src/WpDateTimeZone.php
 	 * @link https://github.com/WordPress/WordPress/blob/4.9.4/wp-includes/functions.php#L72-L151
 	 *
-	 * @return DateTimeZone
+	 * @return \DateTimeZone
 	 */
 	public static function get_default() {
-		$timezone_string = get_option( 'timezone_string' );
+		$timezone_string = \get_option( 'timezone_string' );
 
 		if ( ! empty( $timezone_string ) ) {
 			return new DateTimeZone( $timezone_string );
 		}
 
-		$gmt_offset = get_option( 'gmt_offset' );
+		$gmt_offset = \get_option( 'gmt_offset' );
 		$hours      = (int) $gmt_offset;
-		$minutes    = abs( ( $gmt_offset - (int) $gmt_offset ) * 60 );
-		$offset     = sprintf( '%+03d:%02d', $hours, $minutes );
+		$minutes    = \abs( ( $gmt_offset - (int) $gmt_offset ) * 60 );
+		$offset     = \sprintf( '%+03d:%02d', $hours, $minutes );
 
 		/**
 		 * Offset values as timezone parameter are supported since PHP 5.5.10.
 		 *
 		 * @link http://php.net/manual/en/datetimezone.construct.php
 		 */
-		if ( version_compare( PHP_VERSION, '5.5.10', '<' ) ) {
+		if ( \version_compare( PHP_VERSION, '5.5.10', '<' ) ) {
 			$date = new DateTime( $offset );
 
 			return $date->getTimezone();
 		}
 
-		return new DateTimeZone( $offset );
+		return new \DateTimeZone( $offset );
 	}
 
 	/**
 	 * Get offset.
 	 *
-	 * @param \DateTime $date DateTime object.
-	 *
-	 * @return float|int
+	 * @param \DateTimeInterface $date DateTime object.
+	 * @return int
 	 */
 	public static function get_offset( $date ) {
-		$timezone_string = get_option( 'timezone_string' );
+		$timezone_string = \get_option( 'timezone_string' );
 
 		if ( empty( $timezone_string ) ) {
-			return floatval( get_option( 'gmt_offset', 0 ) ) * HOUR_IN_SECONDS;
+			return \intval( \floatval( \get_option( 'gmt_offset', 0 ) ) * HOUR_IN_SECONDS );
 		}
 
 		$timezone = new DateTimeZone( $timezone_string );