+++ CHANGELOG.md
@@ -7,6 +7,15 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [1.2.0] - 2020-10-08
+- Added DateTimeImmutable class.
+- Added `DateTime::create_from_immutable( \DateTimeImmutable $object )` method.
+- Added `DateTimeImmutable::create_from_mutable( \DateTime $object )` method.
+- Override upstream `DateTime::createFromImmutable( $object )` method.
+- Override upstream `DateTimeImmutable::createFromMutable( $object )` method.
+- Override upstream `DateTimeInterface::createFromFormat( $format, $time, $timezone = null )` method.
+- Updated copyright.
+
 ## [1.1.1] - 2019-12-17
 - Fix for WordPress core trac ticket 48319 (https://core.trac.wordpress.org/ticket/48319).
 - Updated PHP compatibility test version to PHP 5.6.
@@ -26,7 +35,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0
 - First release.
 
-[unreleased]: https://github.com/pronamic/wp-datetime/compare/1.1.1...HEAD
+[unreleased]: https://github.com/pronamic/wp-datetime/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/pronamic/wp-datetime/compare/1.1.1...1.2.0
 [1.1.1]: https://github.com/pronamic/wp-datetime/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/pronamic/wp-datetime/compare/1.0.2...1.1.0
 [1.0.2]: https://github.com/pronamic/wp-datetime/compare/1.0.1...1.0.2